backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Data processing
pandas==2.2.0
openpyxl==3.1.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Code quality
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# AI / Claude API
anthropic==0.18.0
tenacity==8.2.3

# Redis Cache
redis==5.0.0
